+++ codetable/examples/manual_mh.rs
@@ -1,5 +1,5 @@
-use multihash::MultihashDigest;
 use multihash_codetable::Code;
+use multihash_derive::MultihashDigest;
 
 /// prefix/multihash generating tool to aid when adding new tests
 fn prefix_util() {
