+++ derive/Cargo.toml
@@ -8,21 +8,6 @@ license = "MIT"
 repository = "https://github.com/multiformats/multihash"
 resolver = "2"
 
-[lib]
-proc-macro = true
-
 [dependencies]
-proc-macro2 = { version = "1.0.24", features = ["span-locations"] }
-proc-macro-crate = "~1.1.0"
-proc-macro-error = "1.0.4"
-quote = "1.0.7"
-syn = "1.0.42"
-synstructure = "0.12.4"
-
-[features]
-default = ["std"]
-std = []
-
-[dev-dependencies]
-pretty_assertions = "1.0.0"
-#multihash = { path = "..", default-features = false, features = ["derive", "sha2"] }
+multihash-derive-impl = { version = "0.1.0", path = "../derive-impl" }
+multihash = { version = "0.18.0", path = "../" }
